primeria versão de teste VO para usuario

+++ src/main/java/com/sos/sos/entidade/Usuario.java
@@ -31,7 +31,7 @@ public class Usuario {
     private String nome;
 
     @Column(name = "cpf")
-    private Double cpf;
+    private Long cpf;
     
     @Column(name = "UF")
     private String uf;
